requirements.txt
flask
flask-login
flask-sqlalchemy
requests
pycrypto
